pyproject.toml
[tool.poetry]
name = "swarmPlan2Serp"
version = "0.1.0"
description = "a demo for making a AI search engine using swarm and fasthtml"
authors = ["Frank Lin <etrobot@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
duckduckgo-search = "^6.2.11"
python-fasthtml = "^0.5.1"
shortuuid = "^1.0.13"
swarm = { git = "https://git@github.com/openai/swarm.git" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
